requirements.txt
requests
PyYAML
Flask==2.3.2
